requirements.txt
python-telegram-bot
heroku3
docker
python-dotenv